[MODI-64] REFACTOR: merge conflict 해결 - WebSecurityConfiguration 파일 중복의 문제를 해결했습니다.
